// frontend/src/app/api-services/api/model/task-request.ts
/**
 * StudyConnect Backend API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TaskPriority } from './task-priority';
import { TaskStatus } from './task-status';


export interface TaskRequest {
    title: string;
    description?: string;
    due_date?: string;
    priority: TaskPriority;
    status?: TaskStatus;
    category?: string;
    tags?: Array<string>;
    created_by_id: number;
    assignee_ids: Array<number>;
}